backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==1.24.3
prophet==1.1.5
pystan==2.19.1.1
gunicorn==21.2.0
